+++ src/common/input.ts
@@ -1119,12 +1119,12 @@ export class Input extends EventEmitter {
             if (arg.endsWith('\n'))
                 arg = arg.substring(0, args.length - 1);
             if (arg.length > 0) args.push(arg);
-            return this.executeFunction(fun, args, raw);
+            return this.executeFunction(fun, args, raw, cmdChar);
         }
         return txt;
     }
 
-    public executeFunction(fun: string, args, raw: string) {
+    public executeFunction(fun: string, args, raw: string, cmdChar: string) {
         let n;
         let f = false;
         let items;
@@ -1144,7 +1144,7 @@ export class Input extends EventEmitter {
             case 'testfile':
                 args = this.parseInline(args.join(' '));
                 if (!args || args.length === 0)
-                    throw new Error('Invalid syntax use #testfile file');
+                    throw new Error('Invalid syntax use ' + cmdChar + 'testfile file');
                 if (!isFileSync(args))
                     throw new Error('Invalid file "' + args + '"');
                 tmp = fs.readFileSync(args, 'utf-8');
@@ -1160,7 +1160,7 @@ export class Input extends EventEmitter {
                 args = this.parseInline(args.join(' '));
                 items = [];
                 if (!args || args.length === 0)
-                    throw new Error('Invalid syntax use #testspeedfile file');
+                    throw new Error('Invalid syntax use ' + cmdChar + 'testspeedfile file');
                 if (!isFileSync(args))
                     throw new Error('Invalid file "' + args + '"');
                 tmp = fs.readFileSync(args, 'utf-8');
@@ -1190,7 +1190,7 @@ export class Input extends EventEmitter {
                 args = this.parseInline(args.join(' '));
                 items = [];
                 if (!args || args.length === 0)
-                    throw new Error('Invalid syntax use #testspeedfile file');
+                    throw new Error('Invalid syntax use ' + cmdChar + 'testspeedfile file');
                 if (!isFileSync(args))
                     throw new Error('Invalid file "' + args + '"');
                 tmp = fs.readFileSync(args, 'utf-8');
@@ -1234,7 +1234,7 @@ export class Input extends EventEmitter {
                 reload = true;
                 p = path.join(parseTemplate('{data}'), 'profiles');
                 if (args.length < 1 || args.length > 2)
-                    throw new Error('Invalid syntax use \x1b[4m#unt\x1b[0;-11;-12mrigger {pattern|name} \x1b[3mprofile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'unt\x1b[0;-11;-12mrigger {pattern|name} \x1b[3mprofile\x1b[0;-11;-12m');
                 if (args[0].length === 0)
                     throw new Error('Invalid name or pattern');
                 //{pattern} {commands} profile
@@ -1328,7 +1328,7 @@ export class Input extends EventEmitter {
                         }
                         return null;
                     default:
-                        throw new Error('Invalid syntax use \x1b[4m#sus\x1b[0;-11;-12mpend id \x1b[3mprofile\x1b[0;-11;-12m or \x1b[4m#sus\x1b[0;-11;-12mpend');
+                        throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'sus\x1b[0;-11;-12mpend id \x1b[3mprofile\x1b[0;-11;-12m or \x1b[4m' + cmdChar + 'sus\x1b[0;-11;-12mpend');
                 }
             case 'resume':
             case 'resu':
@@ -1353,7 +1353,7 @@ export class Input extends EventEmitter {
                         }
                         return null;
                     default:
-                        throw new Error('Invalid syntax use \x1b[4m#resu\x1b[0;-11;-12mme id \x1b[3mprofile\x1b[0;-11;-12m or \x1b[4m#resu\x1b[0;-11;-12mme');
+                        throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'resu\x1b[0;-11;-12mme id \x1b[3mprofile\x1b[0;-11;-12m or \x1b[4m' + cmdChar + 'resu\x1b[0;-11;-12mme');
                 }
             case 'trigger':
             case 'tr':
@@ -1368,7 +1368,7 @@ export class Input extends EventEmitter {
                 };
                 p = path.join(parseTemplate('{data}'), 'profiles');
                 if (args.length < 2 || args.length > 5)
-                    throw new Error('Invalid syntax use \x1b[4m#tr\x1b[0;-11;-12migger name {pattern} {commands} \x1b[3moptions profile\x1b[0;-11;-12m or \x1b[4m#tr\x1b[0;-11;-12migger {pattern} {commands} \x1b[3m{options} profile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'tr\x1b[0;-11;-12migger name {pattern} {commands} \x1b[3moptions profile\x1b[0;-11;-12m or \x1b[4m' + cmdChar + 'tr\x1b[0;-11;-12migger {pattern} {commands} \x1b[3m{options} profile\x1b[0;-11;-12m');
                 if (args[0].length === 0)
                     throw new Error('Invalid trigger name or pattern');
 
@@ -1483,7 +1483,7 @@ export class Input extends EventEmitter {
                 };
                 p = path.join(parseTemplate('{data}'), 'profiles');
                 if (args.length < 2 || args.length > 4)
-                    throw new Error('Invalid syntax use \x1b[4m#ev\x1b[0;-11;-12ment name {commands} \x1b[3moptions profile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'ev\x1b[0;-11;-12ment name {commands} \x1b[3moptions profile\x1b[0;-11;-12m');
                 if (args[0].length === 0)
                     throw new Error('Invalid event name');
 
@@ -1656,14 +1656,14 @@ export class Input extends EventEmitter {
             case 'une':
                 //#region unevent
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#une\x1b[0;-11;-12mvent name or \x1b[4m#une\x1b[0;-11;-12mvent {name} \x1b[3mprofile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'une\x1b[0;-11;-12mvent name or \x1b[4m' + cmdChar + 'une\x1b[0;-11;-12mvent {name} \x1b[3mprofile\x1b[0;-11;-12m');
                 else {
                     reload = true;
                     profile = null;
                     p = path.join(parseTemplate('{data}'), 'profiles');
                     if (args[0].match(/^\{.*\}$/g) || args[0].match(/^".*"$/g) || args[0].match(/^'.*'$/g)) {
                         if (args.length > 2)
-                            throw new Error('Invalid syntax use \x1b[4m#une\x1b[0;-11;-12mvent name or \x1b[4m#une\x1b[0;-11;-12mvent {name} \x1b[3mprofile\x1b[0;-11;-12m');
+                            throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'une\x1b[0;-11;-12mvent name or \x1b[4m' + cmdChar + 'une\x1b[0;-11;-12mvent {name} \x1b[3mprofile\x1b[0;-11;-12m');
                         if (args.length === 2) {
                             profile = this.parseInline(this.stripQuotes(args[1]));
                             if (this.client.profiles.contains(profile))
@@ -1748,7 +1748,7 @@ export class Input extends EventEmitter {
                 };
                 p = path.join(parseTemplate('{data}'), 'profiles');
                 if (args.length < 2 || args.length > 5)
-                    throw new Error('Invalid syntax use \x1b[4m#bu\x1b[0;-11;-12mtton name|index or \x1b[4m#bu\x1b[0;-11;-12mtton name \x1b[3mcaption\x1b[0;-11;-12m {commands} \x1b[3m{icon} options profile\x1b[0;-11;-12m or \x1b[4m#by\x1b[0;-11;-12mutton \x1b[3mcaption\x1b[0;-11;-12m {commands} \x1b[3m{icon} {options} profile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'bu\x1b[0;-11;-12mtton name|index or \x1b[4m' + cmdChar + 'bu\x1b[0;-11;-12mtton name \x1b[3mcaption\x1b[0;-11;-12m {commands} \x1b[3m{icon} options profile\x1b[0;-11;-12m or \x1b[4m' + cmdChar + 'by\x1b[0;-11;-12mutton \x1b[3mcaption\x1b[0;-11;-12m {commands} \x1b[3m{icon} {options} profile\x1b[0;-11;-12m');
                 if (args[0].length === 0)
                     throw new Error('Invalid button name, caption or commands');
 
@@ -1939,14 +1939,14 @@ export class Input extends EventEmitter {
             case 'unb':
                 //#region unbutton
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#unb\x1b[0;-11;-12mtton name or \x1b[4m#unb\x1b[0;-11;-12mtton {name} \x1b[3mprofile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'unb\x1b[0;-11;-12mtton name or \x1b[4m' + cmdChar + 'unb\x1b[0;-11;-12mtton {name} \x1b[3mprofile\x1b[0;-11;-12m');
                 else {
                     reload = true;
                     profile = null;
                     p = path.join(parseTemplate('{data}'), 'profiles');
                     if (args[0].match(/^\{.*\}$/g) || args[0].match(/^".*"$/g) || args[0].match(/^'.*'$/g)) {
                         if (args.length > 2)
-                            throw new Error('Invalid syntax use \x1b[4m#unb\x1b[0;-11;-12mtton name or \x1b[4m#unb\x1b[0;-11;-12mtton {name} \x1b[3mprofile\x1b[0;-11;-12m');
+                            throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'unb\x1b[0;-11;-12mtton name or \x1b[4m' + cmdChar + 'unb\x1b[0;-11;-12mtton {name} \x1b[3mprofile\x1b[0;-11;-12m');
                         if (args.length === 2) {
                             profile = this.parseInline(this.stripQuotes(args[1]));
                             if (this.client.profiles.contains(profile))
@@ -2016,13 +2016,13 @@ export class Input extends EventEmitter {
                 n = false;
                 p = path.join(parseTemplate('{data}'), 'profiles');
                 if (args.length < 2 || args.length > 4)
-                    throw new Error('Invalid syntax use \x1b[4m#ala\x1b[0;-11;-12mrm name {timepattern} {commands} \x1b[3mprofile\x1b[0;-11;-12m, \x1b[4m#ala\x1b[0;-11;-12mrm name {timepattern} \x1b[3mprofile\x1b[0;-11;-12m, or \x1b[4m#ala\x1b[0;-11;-12mrm {timepattern} {commands} \x1b[3mprofile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'ala\x1b[0;-11;-12mrm name {timepattern} {commands} \x1b[3mprofile\x1b[0;-11;-12m, \x1b[4m' + cmdChar + 'ala\x1b[0;-11;-12mrm name {timepattern} \x1b[3mprofile\x1b[0;-11;-12m, or \x1b[4m' + cmdChar + 'ala\x1b[0;-11;-12mrm {timepattern} {commands} \x1b[3mprofile\x1b[0;-11;-12m');
                 if (args[0].length === 0)
                     throw new Error('Invalid name or timepattern');
                 //{pattern} {commands} profile
                 if (args[0].match(/^\{.*\}$/g)) {
                     if (args.length > 3)
-                        throw new Error('Invalid syntax use \x1b[4m#ala\x1b[0;-11;-12mrm {timepattern} {commands} profile');
+                        throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'ala\x1b[0;-11;-12mrm {timepattern} {commands} profile');
                     args[0] = args[0].substr(1, args[0].length - 2);
                     args[0] = this.parseInline(args[0]);
                     if (args[1].match(/^\{[\s\S]*\}$/g))
@@ -2172,7 +2172,7 @@ export class Input extends EventEmitter {
             case 'ungag':
             case 'ung':
                 if (args.length > 0)
-                    throw new Error('Invalid syntax use \x1b[4m#ung\x1b[0;-11;-12mag number or \x1b[4m#ung\x1b[0;-11;-12mag');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'ung\x1b[0;-11;-12mag number or \x1b[4m' + cmdChar + 'ung\x1b[0;-11;-12mag');
                 if (this._gagID.length) {
                     clearTimeout(this._gagID.pop());
                     this._gags.pop();
@@ -2205,7 +2205,7 @@ export class Input extends EventEmitter {
                     return null;
                 }
                 else if (args.length > 1)
-                    throw new Error('Invalid syntax use \x1b[4m#ga\x1b[0;-11;-12mg number or \x1b[4m#ga\x1b[0;-11;-12mg');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'ga\x1b[0;-11;-12mg number or \x1b[4m' + cmdChar + 'ga\x1b[0;-11;-12mg');
                 i = parseInt(args[0], 10);
                 if (isNaN(i))
                     throw new Error('Invalid number \'' + args[0] + '\'');
@@ -2248,7 +2248,7 @@ export class Input extends EventEmitter {
             case 'wait':
             case 'wa':
                 if (args.length === 0 || args.length > 1)
-                    throw new Error('Invalid syntax use \x1b[4m#wa\x1b[0;-11;-12mit number');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'wa\x1b[0;-11;-12mit number');
                 i = parseInt(this.parseInline(args[0]), 10);
                 if (isNaN(i))
                     throw new Error('Invalid number \'' + i + '\' for wait');
@@ -2282,7 +2282,7 @@ export class Input extends EventEmitter {
                         });
                     });
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use #\x1b[4mraise\x1b[0;-11;-12mevent name or #\x1b[4mraise\x1b[0;-11;-12mevent name arguments');
+                    throw new Error('Invalid syntax use ' + cmdChar + '\x1b[4mraise\x1b[0;-11;-12mevent name or ' + cmdChar + '\x1b[4mraise\x1b[0;-11;-12mevent name arguments');
                 else if (args.length === 1)
                     this.client.raise(args[0]);
                 else
@@ -2303,7 +2303,7 @@ export class Input extends EventEmitter {
                         });
                     });
                 if (args.length === 0 || args.length > 2)
-                    throw new Error('Invalid syntax use #\x1b[4mwin\x1b[0;-11;-12mdow name');
+                    throw new Error('Invalid syntax use ' + cmdChar + '\x1b[4mwin\x1b[0;-11;-12mdow name');
                 else if (args.length === 1)
                     this.client.emit('window', this.stripQuotes(this.parseInline(args[0])));
                 else
@@ -2312,7 +2312,7 @@ export class Input extends EventEmitter {
             case 'raisedelayed':
             case 'raisede':
                 if (args.length < 2)
-                    throw new Error('Invalid syntax use \x1b[4m#raisede\x1b[0;-11;-12mlayed milliseconds name or \x1b[4m#raisede\x1b[0;-11;-12mlayed milliseconds name arguments');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'raisede\x1b[0;-11;-12mlayed milliseconds name or \x1b[4m' + cmdChar + 'raisede\x1b[0;-11;-12mlayed milliseconds name arguments');
                 i = parseInt(this.stripQuotes(args[0]), 10);
                 if (isNaN(i))
                     throw new Error('Invalid number \'' + args[0] + '\' for raisedelayed');
@@ -2340,7 +2340,7 @@ export class Input extends EventEmitter {
             case 'notify':
             case 'not':
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#not\x1b[0;-11;-12mify title \x1b[3mmessage icon\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'not\x1b[0;-11;-12mify title \x1b[3mmessage icon\x1b[0;-11;-12m');
                 else {
                     args[0] = this.stripQuotes(args[0]);
                     if (args[args.length - 1].match(/^\{.*\}$/g)) {
@@ -2348,7 +2348,7 @@ export class Input extends EventEmitter {
                         n = { icon: parseTemplate(this.parseInline(item.substr(1, item.length - 2))) };
                     }
                     if (args.length === 0)
-                        throw new Error('Invalid syntax use \x1b[4m#not\x1b[0;-11;-12mify title \x1b[3mmessage icon\x1b[0;-11;-12m');
+                        throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'not\x1b[0;-11;-12mify title \x1b[3mmessage icon\x1b[0;-11;-12m');
                     if (args.length === 1)
                         this.client.notify(this.parseInline(this.stripQuotes(args[0])), null, n);
                     else
@@ -2480,7 +2480,7 @@ export class Input extends EventEmitter {
             case 'alias':
             case 'al':
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#al\x1b[0;-11;-12mias name value or \x1b[4m#al\x1b[0;-11;-12mias name {value} \x1b[3mprofile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'al\x1b[0;-11;-12mias name value or \x1b[4m' + cmdChar + 'al\x1b[0;-11;-12mias name {value} \x1b[3mprofile\x1b[0;-11;-12m');
                 else if (args.length === 1)
                     throw new Error('Must supply an alias value');
                 else {
@@ -2490,7 +2490,7 @@ export class Input extends EventEmitter {
                     p = path.join(parseTemplate('{data}'), 'profiles');
                     if (args[0].match(/^\{.*\}$/g) || args[0].match(/^".*"$/g) || args[0].match(/^'.*'$/g)) {
                         if (args.length > 2)
-                            throw new Error('Invalid syntax use \x1b[4m#al\x1b[0;-11;-12mias name value or \x1b[4m#al\x1b[0;-11;-12mias name {value} \x1b[3mprofile\x1b[0;-11;-12m');
+                            throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'al\x1b[0;-11;-12mias name value or \x1b[4m' + cmdChar + 'al\x1b[0;-11;-12mias name {value} \x1b[3mprofile\x1b[0;-11;-12m');
                         if (args.length === 2) {
                             profile = this.parseInline(this.stripQuotes(args[1]));
                             if (this.client.profiles.contains(profile))
@@ -2552,14 +2552,14 @@ export class Input extends EventEmitter {
             case 'unalias':
             case 'una':
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#una\x1b[0;-11;-12mlias name or \x1b[4m#una\x1b[0;-11;-12mlias {name} \x1b[3mprofile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'una\x1b[0;-11;-12mlias name or \x1b[4m' + cmdChar + 'una\x1b[0;-11;-12mlias {name} \x1b[3mprofile\x1b[0;-11;-12m');
                 else {
                     reload = true;
                     profile = null;
                     p = path.join(parseTemplate('{data}'), 'profiles');
                     if (args[0].match(/^\{.*\}$/g) || args[0].match(/^".*"$/g) || args[0].match(/^'.*'$/g)) {
                         if (args.length > 2)
-                            throw new Error('Invalid syntax use \x1b[4m#una\x1b[0;-11;-12mlias name or \x1b[4m#una\x1b[0;-11;-12mlias {name} \x1b[3mprofile\x1b[0;-11;-12m');
+                            throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'una\x1b[0;-11;-12mlias name or \x1b[4m' + cmdChar + 'una\x1b[0;-11;-12mlias {name} \x1b[3mprofile\x1b[0;-11;-12m');
                         if (args.length === 2) {
                             profile = this.stripQuotes(args[1]);
                             profile = this.parseInline(profile);
@@ -2621,7 +2621,7 @@ export class Input extends EventEmitter {
             case 'setsetting':
             case 'sets':
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#sets\x1b[0;-11;-12metting name value');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'sets\x1b[0;-11;-12metting name value');
                 else if (args.length === 1)
                     throw new Error('Must supply a setsetting value');
                 else {
@@ -2704,7 +2704,7 @@ export class Input extends EventEmitter {
             case 'getsetting':
             case 'gets':
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#gets\x1b[0;-11;-12metting name');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'gets\x1b[0;-11;-12metting name');
                 else {
                     n = this.stripQuotes(this.parseInline(args.join(' ')));
                     if (/^\s*?\d+\s*?$/.exec(n)) {
@@ -2789,7 +2789,7 @@ export class Input extends EventEmitter {
             case 'profile':
             case 'pro':
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#pro\x1b[0;-11;-12mfile name or \x1b[4m#pro\x1b[0;-11;-12mfile name enable/disable');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'pro\x1b[0;-11;-12mfile name or \x1b[4m' + cmdChar + 'pro\x1b[0;-11;-12mfile name enable/disable');
                 else if (args.length === 1) {
                     args[0] = this.parseInline(args[0]);
                     this.client.toggleProfile(args[0]);
@@ -2810,7 +2810,7 @@ export class Input extends EventEmitter {
                             throw new Error('Profile not found');
                     }
                     if (!args[1])
-                        throw new Error('Invalid syntax use \x1b[4m#pro\x1b[0;-11;-12mfile name or \x1b[4m#pro\x1b[0;-11;-12mfile name enable/disable');
+                        throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'pro\x1b[0;-11;-12mfile name or \x1b[4m' + cmdChar + 'pro\x1b[0;-11;-12mfile name enable/disable');
                     args[1] = this.parseInline(args[1]);
                     switch (args[1].toLowerCase()) {
                         case 'enable':
@@ -2839,7 +2839,7 @@ export class Input extends EventEmitter {
                             }
                             break;
                         default:
-                            throw new Error('Invalid syntax use \x1b[4m#pro\x1b[0;-11;-12mfile name or \x1b[4m#pro\x1b[0;-11;-12mfile name enable/disable');
+                            throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'pro\x1b[0;-11;-12mfile name or \x1b[4m' + cmdChar + 'pro\x1b[0;-11;-12mfile name enable/disable');
                     }
                 }
                 if (this.client.telnet.prompt)
@@ -2862,20 +2862,20 @@ export class Input extends EventEmitter {
                     else
                         item.pattern = this.parseInline(this.stripQuotes(item.pattern));
                     if (args.length === 2) {
-                        item.commands = '#COLOR ' + this.parseInline(args[0]);
+                        item.commands = cmdChar + 'COLOR ' + this.parseInline(args[0]);
                         item.profile = this.stripQuotes(args[1]);
                         if (item.profile.length !== 0)
                             item.profile = this.parseInline(item.profile);
                     }
                     else
-                        item.commands = '#COLOR ' + this.parseInline(args[0]);
+                        item.commands = cmdChar + 'COLOR ' + this.parseInline(args[0]);
                     this.createTrigger(item.pattern, item.commands, item.profile);
                     return null;
                 }
                 else if (args.length !== 1)
-                    throw new Error('Invalid syntax use \x1b[4m#co\x1b[0;-11;-12mlor color or \x1b[4m#co\x1b[0;-11;-12mlor {pattern} color \x1b[3mprofile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'co\x1b[0;-11;-12mlor color or \x1b[4m' + cmdChar + 'co\x1b[0;-11;-12mlor {pattern} color \x1b[3mprofile\x1b[0;-11;-12m');
                 if (args.length !== 1)
-                    throw new Error('Invalid syntax use \x1b[4m#co\x1b[0;-11;-12mlor color or \x1b[4m#co\x1b[0;-11;-12mlor {pattern} color \x1b[3mprofile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'co\x1b[0;-11;-12mlor color or \x1b[4m' + cmdChar + 'co\x1b[0;-11;-12mlor {pattern} color \x1b[3mprofile\x1b[0;-11;-12m');
                 args[0] = this.parseInline(this.stripQuotes(args[0]));
                 n = this.client.display.lines.length;
                 if (args[0].trim().match(/^[-|+]?\d+$/g)) {
@@ -3025,18 +3025,18 @@ export class Input extends EventEmitter {
                     else
                         item.pattern = this.parseInline(this.stripQuotes(item.pattern));
                     if (args.length === 2) {
-                        item.commands = '#CW ' + this.parseInline(args[0]);
+                        item.commands = cmdChar + 'CW ' + this.parseInline(args[0]);
                         item.profile = this.stripQuotes(args[1]);
                         if (item.profile.length !== 0)
                             item.profile = this.parseInline(item.profile);
                     }
                     else
-                        item.commands = '#CW ' + this.parseInline(args[0]);
+                        item.commands = cmdChar + 'CW ' + this.parseInline(args[0]);
                     this.createTrigger(item.pattern, item.commands, item.profile);
                     return null;
                 }
                 else if (args.length !== 1)
-                    throw new Error('Invalid syntax use #cw color or #cw {pattern} color \x1b[3mprofile\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use ' + cmdChar + 'cw color or ' + cmdChar + 'cw {pattern} color \x1b[3mprofile\x1b[0;-11;-12m');
                 //no regex so
                 if (!trigger) return null;
                 args[0] = this.parseInline(this.stripQuotes(args[0]));
@@ -3229,11 +3229,11 @@ export class Input extends EventEmitter {
                 return null;
             case 'pcol':
                 if (args.length < 1 || args.length > 5)
-                    throw new Error('Invalid syntax use #pcol color \x1b[3mXStart, XEnd, YStart, YEnd\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use ' + cmdChar + 'pcol color \x1b[3mXStart, XEnd, YStart, YEnd\x1b[0;-11;-12m');
                 if (args.length > 1) {
                     tmp = [].concat(...args.slice(1).map(s => this.parseInline(this.stripQuotes(s)).split(' ')));
                     if (tmp.length > 4)
-                        throw new Error('Too many arguments use #pcol color \x1b[3mXStart, XEnd, YStart, YEnd\x1b[0;-11;-12m');
+                        throw new Error('Too many arguments use ' + cmdChar + 'pcol color \x1b[3mXStart, XEnd, YStart, YEnd\x1b[0;-11;-12m');
                     item = { xStart: 0 };
                     if (tmp.length > 0)
                         item.xStart = parseInt(tmp[0], 10);
@@ -3382,7 +3382,7 @@ export class Input extends EventEmitter {
                     item = {
                         profile: null,
                         pattern: null,
-                        commands: '#HIGHLIGHT'
+                        commands: cmdChar + 'HIGHLIGHT'
                     };
                     item.pattern = args.shift();
                     if (item.pattern.match(/^\{.*\}$/g))
@@ -3395,7 +3395,7 @@ export class Input extends EventEmitter {
                     return null;
                 }
                 else if (args.length)
-                    throw new Error('Too many arguments use \x1b[4m#hi\x1b[0;-11;-12mghlight \x1b[3mpattern profile\x1b[0;-11;-12m');
+                    throw new Error('Too many arguments use \x1b[4m' + cmdChar + 'hi\x1b[0;-11;-12mghlight \x1b[3mpattern profile\x1b[0;-11;-12m');
                 n = this.client.display.lines.length;
                 setTimeout(() => {
                     n = this.adjustLastLine(n);
@@ -3405,9 +3405,9 @@ export class Input extends EventEmitter {
             case 'break':
             case 'br':
                 if (args.length)
-                    throw new Error('Invalid syntax use \x1b[4m#br\x1b[0;-11;-12meak\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'br\x1b[0;-11;-12meak\x1b[0;-11;-12m');
                 if (!this.loops.length)
-                    throw new Error('\x1b[4m#br\x1b[0;-11;-12meak\x1b[0;-11;-12m must be used in a loop.');
+                    throw new Error('\x1b[4m' + cmdChar + 'br\x1b[0;-11;-12meak\x1b[0;-11;-12m must be used in a loop.');
                 if (this.stack.break)
                     this.stack.break++;
                 else
@@ -3416,14 +3416,14 @@ export class Input extends EventEmitter {
             case 'continue':
             case 'cont':
                 if (args.length)
-                    throw new Error('Invalid syntax use \x1b[4m#cont\x1b[0;-11;-12minue\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'cont\x1b[0;-11;-12minue\x1b[0;-11;-12m');
                 if (!this.loops.length)
-                    throw new Error('\x1b[4m#cont\x1b[0;-11;-12minue\x1b[0;-11;-12m must be used in a loop.');
+                    throw new Error('\x1b[4m' + cmdChar + 'cont\x1b[0;-11;-12minue\x1b[0;-11;-12m must be used in a loop.');
                 this.stack.continue = true;
                 return -2;
             case 'if':
                 if (!args.length || args.length > 3)
-                    throw new Error('Invalid syntax use #if {expression} {true-command} \x1b[3m{false-command}\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use ' + cmdChar + 'if {expression} {true-command} \x1b[3m{false-command}\x1b[0;-11;-12m');
                 if (args[0].match(/^\{[\s\S]*\}$/g))
                     args[0] = args[0].substr(1, args[0].length - 2);
                 tmp = null;
@@ -3443,7 +3443,7 @@ export class Input extends EventEmitter {
             case 'case':
             case 'ca':
                 if (!args.length || args.length < 2)
-                    throw new Error('Invalid syntax use \x1b[4m#ca\x1b[0;-11;-12mse\x1b[0;-11;-12m index {command 1} \x1b[3m{command n}\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'ca\x1b[0;-11;-12mse\x1b[0;-11;-12m index {command 1} \x1b[3m{command n}\x1b[0;-11;-12m');
                 if (args[0].match(/^\{[\s\S]*\}$/g))
                     args[0] = args[0].substr(1, args[0].length - 2);
                 n = this.evaluate(this.parseInline(args[0]));
@@ -3458,7 +3458,7 @@ export class Input extends EventEmitter {
             case 'switch':
             case 'sw':
                 if (!args.length || args.length < 2)
-                    throw new Error('Invalid syntax use \x1b[4m#sw\x1b[0;-11;-12mitch\x1b[0;-11;-12m (expression) {command} \x1b[3m(expression) {command} ... {else_command}\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'sw\x1b[0;-11;-12mitch\x1b[0;-11;-12m (expression) {command} \x1b[3m(expression) {command} ... {else_command}\x1b[0;-11;-12m');
                 if (args.length % 2 === 1)
                     n = args.pop();
                 else
@@ -3488,7 +3488,7 @@ export class Input extends EventEmitter {
             case 'loop':
             case 'loo':
                 if (args.length < 2)
-                    throw new Error('Invalid syntax use \x1b[4m#loo\x1b[0;-11;-12mp\x1b[0;-11;-12m range {commands}');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'loo\x1b[0;-11;-12mp\x1b[0;-11;-12m range {commands}');
                 n = this.parseInline(args.shift()).split(',');
                 args = args.join(' ');
                 if (args.match(/^\{[\s\S]*\}$/g))
@@ -3505,7 +3505,7 @@ export class Input extends EventEmitter {
             case 'repeat':
             case 'rep':
                 if (args.length < 2)
-                    throw new Error('Invalid syntax use \x1b[4m#rep\x1b[0;-11;-12meat\x1b[0;-11;-12m expression {commands}');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'rep\x1b[0;-11;-12meat\x1b[0;-11;-12m expression {commands}');
                 i = args.shift();
                 if (i.match(/^\{[\s\S]*\}$/g))
                     i = i.substr(1, i.length - 2);
@@ -3518,7 +3518,7 @@ export class Input extends EventEmitter {
                 return this.executeForLoop(0, i, args);
             case 'until':
                 if (args.length < 2)
-                    throw new Error('Invalid syntax use #until expression {commands}');
+                    throw new Error('Invalid syntax use ' + cmdChar + 'until expression {commands}');
                 i = args.shift();
                 if (i.match(/^\{[\s\S]*\}$/g))
                     i = i.substr(1, i.length - 2);
@@ -3547,7 +3547,7 @@ export class Input extends EventEmitter {
             case 'while':
             case 'wh':
                 if (args.length < 2)
-                    throw new Error('Invalid syntax use \x1b[4m#wh\x1b[0;-11;-12mile expression {commands}');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'wh\x1b[0;-11;-12mile expression {commands}');
                 i = args.shift();
                 if (i.match(/^\{[\s\S]*\}$/g))
                     i = i.substr(1, i.length - 2);
@@ -3576,7 +3576,7 @@ export class Input extends EventEmitter {
             case 'forall':
             case 'fo':
                 if (args.length < 2)
-                    throw new Error('Invalid syntax use \x1b[4m#fo\x1b[0;-11;-12mrall stringlist {commands}');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'fo\x1b[0;-11;-12mrall stringlist {commands}');
                 i = args.shift();
                 if (i.match(/^\{[\s\S]*\}$/g))
                     i = i.substr(1, i.length - 2);
@@ -3641,7 +3641,7 @@ export class Input extends EventEmitter {
             case 'add':
             case 'ad':
                 if (args.length < 2)
-                    throw new Error('Invalid syntax use \x1b[4m#ad\x1b[0;-11;-12md name value');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'ad\x1b[0;-11;-12md name value');
                 i = args.shift();
                 if (i.match(/^\{[\s\S]*\}$/g))
                     i = i.substr(1, i.length - 2);
@@ -3656,7 +3656,7 @@ export class Input extends EventEmitter {
             case 'math':
             case 'mat':
                 if (args.length < 2)
-                    throw new Error('Invalid syntax use \x1b[4m#mat\x1b[0;-11;-12mh name value');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'mat\x1b[0;-11;-12mh name value');
                 i = args.shift();
                 if (i.match(/^\{[\s\S]*\}$/g))
                     i = i.substr(1, i.length - 2);
@@ -3669,7 +3669,7 @@ export class Input extends EventEmitter {
             case 'evaluate':
             case 'eva':
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#eva\x1b[0;-11;-12mluate expression');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'eva\x1b[0;-11;-12mluate expression');
                 args = '' + this.evaluate(this.parseInline(args.join(' ')));
                 if (this.client.telnet.prompt)
                     this.client.print('\n' + args + '\x1b[0m\n', false);
@@ -3706,12 +3706,12 @@ export class Input extends EventEmitter {
                     }
                 }
                 else if (args.length > 1)
-                    throw new Error('Invalid syntax use \x1b[4m#fr\x1b[0;-11;-12mEEZE \x1b[3mnumber\x1b[0;-11;-12m');
+                    throw new Error('Invalid syntax use \x1b[4m' + cmdChar + 'fr\x1b[0;-11;-12mEEZE \x1b[3mnumber\x1b[0;-11;-12m');
                 return null;
             //#endregion freeze                
             case 'clr':
                 if (args.length)
-                    throw new Error('Invalid syntax use #CLR');
+                    throw new Error('Invalid syntax use ' + cmdChar + 'CLR');
                 //nothing to clear so just bail
                 if (this.client.display.lines.length === 0)
                     return null;
@@ -3732,7 +3732,7 @@ export class Input extends EventEmitter {
         if (fun.match(/^[-|+]?\d+$/)) {
             i = parseInt(fun, 10);
             if (args.length === 0)
-                throw new Error('Invalid syntax use #nnn commands');
+                throw new Error('Invalid syntax use ' + cmdChar + 'nnn commands');
             args = args.join(' ');
             if (args.match(/^\{[\s\S]*\}$/g))
                 args = args.substr(1, args.length - 2);
@@ -6032,8 +6032,9 @@ export class Input extends EventEmitter {
         const ll = lines.length;
         const code = [];
         const b = [];
+        const cmdChar = this.client.options.commandChar;
         for (; l < ll; l++) {
-            if (lines[l].trim().startsWith('#wait ')) {
+            if (lines[l].trim().startsWith(cmdChar + 'wait ')) {
                 code.push('setTimeout(()=> {');
                 b.unshift(parseInt(lines[l].trim().substr(5), 10) || 0);
             }